Move config validations to own file
